requirements.txt
Django==5.2.4
django-ninja==1.1.0
Pillow==10.1.0
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework==3.14.0